Added test: should be a GET operation when useGETForQueries=true
