+++ packages/executors/http/tests/handleEventStreamResponse.test.ts
@@ -1,34 +1,5 @@
 import { handleEventStreamResponse } from '../src/handleEventStreamResponse.js';
-import { buildHTTPExecutor, SyncFetchFn } from '../src/index.js';
 import { ReadableStream, Response } from '@whatwg-node/fetch';
-import { parse } from 'graphql';
-
-describe('buildHTTPExecutor', () => {
-  it('should be a GET operation when useGETForQueries=true', async () => {
-    const syncFetch: SyncFetchFn = (input, init) => {
-      return {
-        ...new Response(),
-        url: input,
-        headers: new Headers(init?.headers),
-        text: () => JSON.stringify({ data: init }),
-      };
-    };
-
-    const executor = buildHTTPExecutor({
-      useGETForQueries: true,
-      fetch: syncFetch,
-    });
-
-    const mutation = parse(/* GraphQL */ `
-      mutation {
-        doSomething
-      }
-    `);
-
-    const res = executor({ document: mutation });
-    expect(res.data.method).toBe('POST');
-  });
-});
 
 describe('handleEventStreamResponse', () => {
   const encoder = new TextEncoder();
